feature: generic bluetooth app
extract services from MainViewModel.kt
