+++ app/src/main/java/com/milen/GenericBluetoothApp.kt
@@ -2,5 +2,4 @@ package com.milen
 
 import android.app.Application
 
-class GenericBluetoothApp : Application() {
-}
+class GenericBluetoothApp : Application()
